+++ src/Api/Extensions/ConfigurationSetup.cs
@@ -1,0 +1,14 @@
+using SharedKernel;
+
+namespace Api.Extensions;
+
+public static class ConfigurationSetup
+{
+    public static void AddConfigurations(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddOptions<AppSettings>()
+            .Bind(configuration)
+            .ValidateDataAnnotations()
+            .ValidateOnStart();
+    }
+}