+++ src/Api/Program.cs
@@ -1,5 +1,5 @@
 using Api.Extensions;
-using Api.Extensions.Messaging;
+using Api.Extensions.CustomProblemDetails;
 using Api.Extensions.Swagger;
 using Api.Extensions.Telemetry;
 using Thiagosza.Mediator.Core.Extensions;
@@ -7,21 +7,24 @@ using Thiagosza.Mediator.Core.Extensions;
 var builder = WebApplication.CreateBuilder(args);
 Console.WriteLine($"Current environment: {builder.Environment.EnvironmentName}");
 
+builder.Services.AddConfigurations(builder.Configuration);
 builder.Services.AddVersioning();
 builder.Services.AddCors();
 builder.Services.AddApiProblemDetails();
 builder.Services.AdicionaSwagger();
-builder.Services.AddAuthentication(builder.Configuration);
+builder.Services.AddApiAuthentication();
 builder.Services.AddAuthorization();
-builder.Services.RegisterServices(builder.Configuration);
+builder.Services.RegisterServices();
 builder.Services.AddMediator();
 builder.Services.AddValidators();
-builder.Services.AddTelemetry(builder.Host, builder.Configuration);
-builder.Services.AddRabbitMq(builder.Configuration);
+builder.Services.AddTelemetry(builder.Host);
+builder.Services.AddRabbitMq();
 
 var app = builder.Build();
 app.MapEndpoints();
 app.ConfigureSwagger();
 app.UseAuthentication();
 app.UseAuthorization();
+app.UseExceptionHandler();
+app.UseStatusCodePages();
 app.Run();